EMI: First attempt at lighting.
This adds support for ambient, direct and omnidirectional lights. Lighting is done by calculating vertex colors in software, since the attenuation model used in EMI cannot be simulated with GL fixed function lighting. The original game does the lighting this way as well. A shader would be a more modern approach for this.